Run randomx-tests too

+++ build-librandomx.sh
@@ -20,7 +20,7 @@ Darwin)
   ;;
 esac
 destination="$(dirname "$0")/native/$shared_lib_filename"
-if [[ -f "$destination" ]]; then
+if [[ -f "$destination" && "$CI" != "true" ]]; then
   echo "$destination already exists; this script will do nothing." > /dev/stderr
   exit 0
 fi
@@ -37,6 +37,18 @@ git clone \
   git://github.com/tevador/RandomX.git \
   "$tmpdir"
 pushd "$tmpdir"
+if command -v clang > /dev/null; then
+  export CC=clang
+  export CXX=clang++
+fi
+if [[ "$CI" = "true" ]]; then
+  cmake -DARCH=native
+  make
+  ./randomx-tests
+  $CC -o api-example -lm -lstdc++ src/tests/api-example1.c librandomx.a
+  ./api-example
+  make clean
+fi
 cmake -DARCH=native -DBUILD_SHARED_LIBS=ON
 make
 popd
